% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_project_manager}
\alias{addin_project_manager}
\title{Addin para criar e trocar rapidamente de projeto}
\usage{
addin_project_manager()
}
\description{
Addin para criar e trocar rapidamente de projeto
}
