% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{code_format}
\alias{code_format}
\title{Formata um arquivo suportado com styler}
\usage{
code_format(path = NULL)
}
\arguments{
\item{path}{Caminho do arquivo. Quando \code{NULL}, usa o arquivo ativo no editor.}
}
\value{
Invisivelmente, uma lista com o resultado da formatacao.
}
\description{
Formata um arquivo suportado com styler
}
