% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{code_format_all}
\alias{code_format_all}
\title{Formata os arquivos suportados de um projeto}
\usage{
code_format_all(path = ".")
}
\arguments{
\item{path}{Caminho da pasta do projeto.}
}
\value{
Invisivelmente, uma lista com os arquivos formatados.
}
\description{
Formata os arquivos suportados de um projeto
}
