% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{discard_file_changes}
\alias{discard_file_changes}
\title{Descarta mudancas locais de arquivos}
\usage{
discard_file_changes(files, path = ".")
}
\arguments{
\item{files}{Caminhos dos arquivos relativos ao repositorio.}

\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, uma lista com o resultado.
}
\description{
Descarta mudancas locais de arquivos
}
