% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_delete}
\alias{file_delete}
\title{Deleta um arquivo ou pasta no projeto}
\usage{
file_delete(path_to_delete, path = ".", remove_from_git = FALSE)
}
\arguments{
\item{path_to_delete}{Caminho relativo do arquivo/pasta a deletar (ex: "scripts/old_analysis.R")}

\item{path}{Caminho do projeto}

\item{remove_from_git}{Se \code{TRUE}, prepara a remocao no Git quando o item era rastreado.}
}
\value{
Lista com resultado da delecao
}
\description{
Deleta um arquivo ou pasta no projeto
}
