% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_rename}
\alias{file_rename}
\title{Renomeia um arquivo ou pasta no projeto}
\usage{
file_rename(source, target, path = ".")
}
\arguments{
\item{source}{Caminho do arquivo ou pasta atual, relativo ao projeto.}

\item{target}{Novo caminho (nome) do arquivo ou pasta, relativo ao projeto.}

\item{path}{Caminho do projeto. Por padrao, o diretorio de trabalho atual.}
}
\value{
Invisivelmente, uma lista com o resultado da operacao.
}
\description{
Renomeia um arquivo ou pasta no projeto
}
