% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{find_recent_downloads}
\alias{find_recent_downloads}
\title{Encontra os arquivos de dados e scripts mais recentes na pasta Downloads}
\usage{
find_recent_downloads()
}
\value{
Um vetor de caminhos de arquivos.
}
\description{
Encontra os arquivos de dados e scripts mais recentes na pasta Downloads
}
