% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{format_project_files}
\alias{format_project_files}
\title{Formata os arquivos suportados de um projeto}
\usage{
format_project_files(path = ".")
}
\arguments{
\item{path}{Caminho da pasta do projeto.}
}
\value{
Invisivelmente, uma lista com os arquivos formatados.
}
\description{
Formata os arquivos suportados de um projeto
}
