% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{git4stats-package}
\alias{git4stats}
\alias{git4stats-package}
\title{Utilidades internas do pacote git4stats}
\description{
Funcoes auxiliares para diagnostico, mensagens e interacao segura com Git.
}
\author{
\strong{Maintainer}: OpenAI Codex \email{support@openai.com}

Authors:
\itemize{
  \item OpenAI Codex \email{support@openai.com}
}

}
\keyword{internal}
