% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{git_changed_files}
\alias{git_changed_files}
\title{Lista arquivos com mudancas no projeto}
\usage{
git_changed_files(path = ".")
}
\arguments{
\item{path}{Caminho do projeto.}
}
\value{
Um data frame com arquivos alterados.
}
\description{
Lista arquivos com mudancas no projeto
}
