% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{git_clone_repo}
\alias{git_clone_repo}
\title{Clona um repositório Git em uma pasta local}
\usage{
git_clone_repo(remote_url, path = ".", directory = NULL, open = FALSE)
}
\arguments{
\item{remote_url}{URL ou caminho do repositório remoto.}

\item{path}{Pasta base onde o clone será criado.}

\item{directory}{Nome opcional da pasta de destino.}

\item{open}{Se \code{TRUE}, tenta abrir o projeto clonado no RStudio ao final.}
}
\value{
Uma lista com o resultado da operação.
}
\description{
Clona um repositório Git em uma pasta local
}
