% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{git_commit}
\alias{git_commit}
\title{Cria commit com arquivos ja preparados}
\usage{
git_commit(message = "Atualiza projeto", path = ".")
}
\arguments{
\item{message}{Mensagem do commit.}

\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, uma lista com o resultado.
}
\description{
Cria commit com arquivos ja preparados
}
