% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_commit.R
\name{git_commit_all}
\alias{git_commit_all}
\title{Faz o primeiro commit do projeto}
\usage{
git_commit_all(message = "Primeiro commit", path = ".")
}
\arguments{
\item{message}{Mensagem do commit.}

\item{path}{Caminho do projeto.}
}
\value{
Uma lista com o resultado da operacao.
}
\description{
Adiciona os arquivos do projeto ao historico local e cria um commit inicial.
}
