% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{git_diff}
\alias{git_diff}
\title{Mostra o diff de um arquivo do projeto}
\usage{
git_diff(file, path = ".", staged = FALSE, context = c("changes", "full"))
}
\arguments{
\item{file}{Caminho do arquivo relativo ao repositorio.}

\item{path}{Caminho do projeto.}

\item{staged}{Se \code{TRUE}, mostra mudancas preparadas para commit.}

\item{context}{Use \code{"changes"} para mostrar apenas os trechos alterados ou
\code{"full"} para mostrar o arquivo com mais contexto.}
}
\value{
Invisivelmente, uma lista com o diff.
}
\description{
Mostra o diff de um arquivo do projeto
}
