% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{git_discard}
\alias{git_discard}
\title{Descarta mudancas locais de arquivos}
\usage{
git_discard(files, path = ".")
}
\arguments{
\item{files}{Caminhos dos arquivos relativos ao repositorio.}

\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, uma lista com o resultado.
}
\description{
Descarta mudancas locais de arquivos
}
