% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{git_fetch}
\alias{git_fetch}
\title{Busca atualizações do remote sem alterar a branch local}
\usage{
git_fetch(path = ".", remote = "origin")
}
\arguments{
\item{path}{Caminho do projeto.}

\item{remote}{Nome do remote.}
}
\value{
Uma lista com o resultado da operação.
}
\description{
Executa \verb{git fetch --prune} para atualizar as referências remotas.
}
