% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_actions.R
\name{git_set_identity}
\alias{git_set_identity}
\title{Configura nome e email global no Git}
\usage{
git_set_identity(name, email)
}
\arguments{
\item{name}{Nome a registrar em \code{user.name} na configuracao global do Git.}

\item{email}{Email a registrar em \code{user.email} na configuracao global do Git.}
}
\value{
Invisivelmente, \code{TRUE} se a configuracao foi aplicada, \code{FALSE} caso
contrario.
}
\description{
Configura nome e email global no Git
}
