% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{git_setup_wizard}
\alias{git_setup_wizard}
\title{Wizard visual para configurar Git em um projeto RStudio}
\usage{
git_setup_wizard(path = ".")
}
\arguments{
\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, o caminho analisado.
}
\description{
Wizard visual para configurar Git em um projeto RStudio
}
