% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{git_status_pretty}
\alias{git_status_pretty}
\title{Mostra o estado Git do projeto em linguagem amigavel}
\usage{
git_status_pretty(path = ".")
}
\arguments{
\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, um data frame com o status do repositorio.
}
\description{
Mostra o estado Git do projeto em linguagem amigavel
}
