% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{git_unstage}
\alias{git_unstage}
\title{Remove arquivos da preparacao para commit}
\usage{
git_unstage(files, path = ".")
}
\arguments{
\item{files}{Caminhos dos arquivos relativos ao repositorio.}

\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, uma lista com o resultado.
}
\description{
Remove arquivos da preparacao para commit
}
