% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manager.R
\name{project_find}
\alias{project_find}
\title{Lista projetos RStudio disponiveis em uma pasta}
\usage{
project_find(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{Pasta base da busca.}

\item{recursive}{Se \code{TRUE}, procura tambem em subpastas.}
}
\value{
Um data frame com nome, pasta e arquivo \code{.Rproj}.
}
\description{
Lista projetos RStudio disponiveis em uma pasta
}
