% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manager.R
\name{project_manager_addin}
\alias{project_manager_addin}
\title{Addin visual para criar e abrir projetos RStudio}
\usage{
project_manager_addin()
}
\value{
Invisivelmente, o caminho do projeto criado ou aberto.
}
\description{
Addin visual para criar e abrir projetos RStudio
}
