% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manager.R
\name{project_open}
\alias{project_open}
\title{Abre rapidamente um projeto RStudio}
\usage{
project_open(path)
}
\arguments{
\item{path}{Caminho de uma pasta de projeto ou do proprio arquivo \code{.Rproj}.}
}
\value{
Uma lista com o resultado da operacao.
}
\description{
Abre rapidamente um projeto RStudio
}
