% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{stage_files}
\alias{stage_files}
\title{Prepara arquivos para o proximo commit}
\usage{
stage_files(files, path = ".")
}
\arguments{
\item{files}{Caminhos dos arquivos relativos ao repositorio.}

\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, uma lista com o resultado.
}
\description{
Prepara arquivos para o proximo commit
}
