% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{statgit-package}
\alias{statgit-package}
\title{Utilidades internas do pacote statgit}
\description{
Funcoes auxiliares para diagnostico, mensagens e interacao segura com Git.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geovani-ariel/statgit}
  \item Report bugs at \url{https://github.com/geovani-ariel/statgit/issues}
}

}
\author{
\strong{Maintainer}: Geovani Ariel Bueno Paschoini \email{geovaniariel08@hotmail.com}

Authors:
\itemize{
  \item Geovani Ariel Bueno Paschoini \email{geovaniariel08@hotmail.com}
}

}
\keyword{internal}
