% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_panel.R
\name{statgit}
\alias{statgit}
\title{Abre o painel principal do statgit}
\usage{
statgit(path = active_project_path())
}
\arguments{
\item{path}{Caminho do projeto.}
}
\value{
Invisivelmente, o caminho analisado.
}
\description{
Centraliza as acoes do pacote em uma interface Shiny com navegacao por
modulos.
}
