% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{git_clone_repo}
\alias{git_clone_repo}
\title{Clone a Git repository into a local directory}
\usage{
git_clone_repo(remote_url, path = ".", directory = NULL, open = FALSE)
}
\arguments{
\item{remote_url}{URL or path of the remote repository.}

\item{path}{Base directory where the clone will be created.}

\item{directory}{Optional target directory name.}

\item{open}{If \code{TRUE}, tries to open the cloned project in RStudio at the end.}
}
\value{
A list with the operation result.
}
\description{
Clone a Git repository into a local directory
}
