% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{git_fetch}
\alias{git_fetch}
\title{Fetch remote updates without changing the local branch}
\usage{
git_fetch(path = ".", remote = "origin")
}
\arguments{
\item{path}{Project path.}

\item{remote}{Remote name.}
}
\value{
A list with the operation result.
}
\description{
Runs \verb{git fetch --prune} to update remote references.
}
