% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_open_repo}
\alias{github_open_repo}
\title{Abre o repositório GitHub configurado no navegador}
\usage{
github_open_repo(path = ".", remote = "origin")
}
\arguments{
\item{path}{Caminho do projeto.}

\item{remote}{Nome do remote.}
}
\value{
Uma lista com o resultado da operação.
}
\description{
Converte a URL do remote em uma URL web navegável e abre no navegador padrão.
}
